model_definitions/99244ab1-82fb-4bf8-973f-c595cbe09ef4/model_modules/requirements.txt
lightgbm==2.3.0
pypmml==0.9.7
sklearn-pandas==2.0.0
sklearn-pmml-model==0.0.15
sklearn2pmml==0.61.0
